+++ demos/MAUITodo/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿namespace MAUITodo;
+
+public partial class App : Application
+{
+	public App()
+	{
+		InitializeComponent();
+
+		MainPage = new AppShell();
+	}
+}