+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -1,3 +1,8 @@
+# PowerSync.Common Changelog
+
+## 0.0.3-alpha.1
+- Minor changes to accommodate PowerSync.MAUI package extension.
+
 ## 0.0.2-alpha.2
 
 - Updated core extension to v0.3.14